requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic[email]
# testing libs
pytest
httpx
python-jose[cryptography]
ruff
python-multipart
httpx